agent-service/requirements.txt
google-adk
Flask
psycopg2-binary
requests
google-generativeai